+++ int_test/mysql/int_test.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+set -e
+
 # 
 
 if [ "$(basename "${PWD}")" == 'mysql' ]; then