+++ int_test/postgres/int_test.sh
@@ -12,7 +12,7 @@ fi
 dc="int_test/postgres/docker-compose.yml"
 
 # tear down old tests
-docker-compose -f "$dc" down -v
+docker-compose -f "$dc" rm -s -f -v data_check
 
 docker-compose -f "$dc" build --no-rm
 docker-compose -f "$dc" up -d db