+++ int_test/scripts/run_tests.sh
@@ -3,7 +3,7 @@
 DB="$1"
 
 # wait for DB to start
-while ! data_check --config "int_test/${DB}/data_check.yml" --ping; do
+while ! poetry run data_check --config "int_test/${DB}/data_check.yml" --ping; do
     echo "waiting for db"
     sleep 1
 done
@@ -11,28 +11,28 @@ done
 result=0
 
 # run basic unit tests first
-if ! pytest test; then
+if ! poetry run pytest test; then
     result=1
 fi
 
 # then the integration tests
 cd "int_test/${DB}" || exit 1
-if ! pytest ../../test; then
+if ! poetry run pytest ../../test; then
     result=1
 fi
 
 # and start data_check
-if ! data_check --gen; then
+if ! poetry run data_check --gen; then
     result=1
 fi
 
 echo "testing posivite run"
-if ! data_check checks/basic checks/generated --traceback; then
+if ! poetry run data_check checks/basic checks/generated --traceback; then
     result=1
 fi
 
 echo "testing negative run"
-if data_check checks/failing; then
+if poetry run data_check checks/failing; then
     result=1
 fi