+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - --verbose argument
 - --traceback argument
 - -g argument as an alias for --generate
+- colored output
 
 ## [0.1.1] - 2021-02-07